--- src/components/general/Nav.jsx
@@ -1,333 +1,0 @@
-import React, { useContext } from "react";
-import { useState } from "react";
-import { Link, useNavigate } from "react-router-dom";
-import { AuthContext } from "../../context/AuthContext";
-import { googleLogout } from "@react-oauth/google";
-import logoImg from "../../static/assets/png/LOGO2.png";
-import "@animxyz/core";
-import { XyzTransition } from "@animxyz/react";
-export default function Nav() {
-  const { currentUser } = useContext(AuthContext);
-  const { dispatch } = useContext(AuthContext);
-  const nevigate = useNavigate();
-  const [mobileMenu, setMobileMenu] = useState(false);
-  const [profileMenu, setProfileMenu] = useState(false);
-  const [searchMenu, setSearchMenu] = useState(false);
-
-  const logout = () => {
-    googleLogout();
-    dispatch({ type: "LOGOUT" });
-    localStorage.clear();
-    nevigate("/");
-  };
-  return (
-    <nav className="fixed bg-gray-800 top-0 w-full z-50">
-      <div className="max-w-7xl mx-auto px-2 sm:px-6 lg:px-8">
-        <div className="relative flex items-center justify-between h-16">
-          <div className="absolute inset-y-0 left-0 flex items-center sm:hidden">
-            <button
-              type="button"
-              className="inline-flex items-center justify-center p-2 rounded-md text-gray-400 hover:text-white hover:bg-gray-700 focus:outline-none focus:ring-2 focus:ring-inset focus:ring-white"
-              aria-controls="mobile-menu"
-              aria-expanded="false"
-              onClick={() => setMobileMenu(!mobileMenu)}
-            >
-              <span className="sr-only">Open main menu</span>
-
-              <svg
-                xmlns="http://www.w3.org/2000/svg"
-                className="h-6 w-6 block"
-                fill="none"
-                viewBox="0 0 24 24"
-                stroke="currentColor"
-                strokeWidth={2}
-              >
-                <path
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                  d="M4 6h16M4 12h16M4 18h16"
-                />
-              </svg>
-            </button>
-          </div>
-          <div className="flex-1 flex items-center justify-center sm:items-stretch sm:justify-start">
-            <div className="flex-shrink-0 flex items-center">
-              <img
-                className="block lg:hidden h-14 w-auto"
-                src={logoImg}
-                alt="Workflow"
-              />
-              <img
-                className="hidden lg:block h-14 w-auto"
-                src={logoImg}
-                alt="Workflow"
-              />
-            </div>
-            <div className="hidden sm:block sm:ml-6">
-              <div className="flex h-full items-center space-x-4">
-                <Link
-                  to="/"
-                  className="text-gray-300 hover:bg-gray-700 hover:text-white px-3 py-2 rounded-md text-sm font-medium"
-                >
-                  Home
-                </Link>
-
-                <Link
-                  to="/news"
-                  className="text-gray-300 hover:bg-gray-700 hover:text-white px-3 py-2 rounded-md text-sm font-medium"
-                >
-                  News
-                </Link>
-
-                <Link
-                  to="forum"
-                  className="text-gray-300 hover:bg-gray-700 hover:text-white px-3 py-2 rounded-md text-sm font-medium"
-                >
-                  Forum
-                </Link>
-              </div>
-            </div>
-          </div>
-          {currentUser && (
-            <div className="absolute inset-y-0 right-0 flex items-center pr-2 sm:static sm:inset-auto sm:ml-6 sm:pr-0">
-              <div className="mr-3 relative sm:block hidden">
-                <svg
-                  xmlns="http://www.w3.org/2000/svg"
-                  className="h-6 w-6 text-gray-400 hover:text-white cursor-pointer"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                  stroke="currentColor"
-                  strokeWidth={2}
-                  onClick={() => setSearchMenu(true)}
-                >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    d="M21 21l-6-6m2-5a7 7 0 11-14 0 7 7 0 0114 0z"
-                  />
-                </svg>
-              </div>
-              <button
-                type="button"
-                className="bg-gray-800 p-1 rounded-full text-gray-400 hover:text-white focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-offset-gray-800 focus:ring-white"
-              >
-                <span className="sr-only">View notifications</span>
-
-                <svg
-                  xmlns="http://www.w3.org/2000/svg"
-                  className="h-6 w-6"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                  stroke="currentColor"
-                  strokeWidth={2}
-                >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    d="M15 17h5l-1.405-1.405A2.032 2.032 0 0118 14.158V11a6.002 6.002 0 00-4-5.659V5a2 2 0 10-4 0v.341C7.67 6.165 6 8.388 6 11v3.159c0 .538-.214 1.055-.595 1.436L4 17h5m6 0v1a3 3 0 11-6 0v-1m6 0H9"
-                  />
-                </svg>
-              </button>
-
-              <div className="ml-3 relative">
-                <div>
-                  <button
-                    type="button"
-                    className="bg-gray-800 flex text-sm rounded-full focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-offset-gray-800 focus:ring-white"
-                    id="user-menu-button"
-                    aria-expanded="false"
-                    aria-haspopup="true"
-                    onClick={() => setProfileMenu(!profileMenu)}
-                  >
-                    <span className="sr-only">Open user menu</span>
-                    {currentUser.profilePic && (
-                      <img
-                        className="h-8 w-8 rounded-full"
-                        src={currentUser.profilePic}
-                        alt={currentUser.name}
-                      />
-                    )}
-                    {!currentUser.profilePic && (
-                      <h2 className="uppercase w-8 h-8 text-2xl rounded-full flex justify-center items-center">
-                        {currentUser.email.charAt(0)}
-                      </h2>
-                    )}
-                  </button>
-                </div>
-                {profileMenu && (
-                  <div
-                    className="origin-top-right absolute right-0 mt-2 w-48 rounded-md shadow-lg py-1 bg-white ring-1 ring-black ring-opacity-5 focus:outline-none"
-                    role="menu"
-                    aria-orientation="vertical"
-                    aria-labelledby="user-menu-button"
-                    tabIndex="-1"
-                  >
-                    <Link
-                      to="#"
-                      className="block px-4 py-2 text-sm text-gray-700"
-                      role="menuitem"
-                      tabIndex="-1"
-                      id="user-menu-item-0"
-                    >
-                      Your Profile
-                    </Link>
-                    <Link
-                      to="#"
-                      className="block px-4 py-2 text-sm text-gray-700"
-                      role="menuitem"
-                      tabIndex="-1"
-                      id="user-menu-item-1"
-                    >
-                      Settings
-                    </Link>
-                    <button
-                      onClick={logout}
-                      className="block px-4 py-2 text-sm text-gray-700"
-                    >
-                      Sign out
-                    </button>
-                  </div>
-                )}
-              </div>
-            </div>
-          )}
-          {!currentUser && (
-            <div className="absolute  inset-y-0 right-0 flex items-center pr-2 sm:static sm:inset-auto sm:ml-6 sm:pr-0">
-              <div className="mr-3 relative sm:block hidden">
-                <svg
-                  xmlns="http://www.w3.org/2000/svg"
-                  className="h-6 w-6 text-gray-400 hover:text-white cursor-pointer"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                  stroke="currentColor"
-                  strokeWidth={2}
-                  onClick={() => setSearchMenu(true)}
-                >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    d="M21 21l-6-6m2-5a7 7 0 11-14 0 7 7 0 0114 0z"
-                  />
-                </svg>
-              </div>
-              <Link
-                to="/login"
-                className="bg-gray-800 p-1 flex  text-gray-400 hover:text-white focus:outline-none "
-              >
-                <span className="sr-only">Login</span>
-                <svg
-                  xmlns="http://www.w3.org/2000/svg"
-                  className="h-6 w-6 mr-3"
-                  fill="none"
-                  viewBox="0 0 24 24"
-                  stroke="currentColor"
-                  strokeWidth={2}
-                >
-                  <path
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    d="M11 16l-4-4m0 0l4-4m-4 4h14m-5 4v1a3 3 0 01-3 3H6a3 3 0 01-3-3V7a3 3 0 013-3h7a3 3 0 013 3v1"
-                  />
-                </svg>
-                Login
-              </Link>
-            </div>
-          )}
-        </div>
-      </div>
-      <XyzTransition appear xyz="fade-75% up-2  ease-out-back">
-        {mobileMenu && (
-          <div className="sm:hidden px-4" id="mobile-menu">
-            <div className="flex border-b  max-w-lg w-full mx-auto border-white justify-between items-center">
-              <input
-                type="text"
-                onChange={(e) => console.log(e.target.value)}
-                className="w-full capitalize pr-2 py-1 focus:outline-none bg-transparent text-white"
-                placeholder="search post...."
-              />
-              <svg
-                xmlns="http://www.w3.org/2000/svg"
-                className="h-6 w-6 text-white cursor-pointer"
-                fill="none"
-                viewBox="0 0 24 24"
-                stroke="currentColor"
-              >
-                <path
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                  strokeWidth="2"
-                  d="M21 21l-6-6m2-5a7 7 0 11-14 0 7 7 0 0114 0z"
-                />
-              </svg>
-            </div>
-            <div className="px-2 pt-2 pb-3 space-y-1">
-              <Link
-                to="/"
-                className="text-gray-300 hover:bg-gray-700 hover:text-white block px-3 py-2 rounded-md text-base font-medium"
-              >
-                Home
-              </Link>
-
-              <Link
-                to="/news"
-                className="text-gray-300 hover:bg-gray-700 hover:text-white block px-3 py-2 rounded-md text-base font-medium"
-              >
-                News
-              </Link>
-
-              <Link
-                to="/forum"
-                className="text-gray-300 hover:bg-gray-700 hover:text-white block px-3 py-2 rounded-md text-base font-medium"
-              >
-                Forum
-              </Link>
-            </div>
-          </div>
-        )}
-      </XyzTransition>
-      <XyzTransition appear xyz="fade-75% up-2  ease-out-back">
-        {searchMenu && (
-          <div className="w-full backdrop-blur-sm absolute top-0 left-0 mx-auto bg-black bg-opacity-70 h-screen flex justify-center items-center flex-col text-white">
-            <svg
-              xmlns="http://www.w3.org/2000/svg"
-              className="h-6 w-6 absolute top-10 right-10 cursor-pointer"
-              fill="none"
-              viewBox="0 0 24 24"
-              stroke="currentColor"
-              strokeWidth={2}
-              onClick={() => setSearchMenu(false)}
-            >
-              <path
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                d="M6 18L18 6M6 6l12 12"
-              />
-            </svg>
-            <div className="flex border-b-2  max-w-lg w-full mx-auto border-white justify-between items-center">
-              <input
-                type="text"
-                onChange={(e) => console.log(e.target.value)}
-                className="w-full capitalize placeholder:text-white pr-2 py-1 focus:outline-none bg-transparent text-white"
-                placeholder="search post...."
-              />
-              <svg
-                xmlns="http://www.w3.org/2000/svg"
-                className="h-6 w-6 text-white cursor-pointer"
-                fill="none"
-                viewBox="0 0 24 24"
-                stroke="currentColor"
-              >
-                <path
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                  strokeWidth="2"
-                  d="M21 21l-6-6m2-5a7 7 0 11-14 0 7 7 0 0114 0z"
-                />
-              </svg>
-            </div>
-          </div>
-        )}
-      </XyzTransition>
-    </nav>
-  );
-}
