+++ src/components/general/ScrollToTop.jsx
@@ -1,4 +1,3 @@
-import React from "react";
 import { useEffect } from "react";
 import { useLocation } from "react-router-dom";
